CA, OK, AZ Spark modeling (recidiviz/recidiviz-data#6321)
* feb 22 spark policies

* alternate outflows policy function test

* pylint fix

* black

* projection type bug-fix

* zack nit

* lint

GitOrigin-RevId: 7beb09c81f6f090a9ba2bc0913151f53702042e6

+++ recidiviz/calculator/modeling/population_projection/spark_compartment.py
@@ -64,7 +64,7 @@ class SparkCompartment(ABC):
             )
 
     @abstractmethod
-    def ingest_incoming_cohort(self, influx: Dict[str, int]):
+    def ingest_incoming_cohort(self, influx: Dict[str, float]):
         """Ingest the population coming from one compartment into another by the end of the `current_ts`
 
         influx: dictionary of cohort type (str) to number of people revoked for the time period (int)
